refactor(repr): make Row's Columnar Ref a DatumSeq, not &RowRef
Move DatumSeq from mz_compute::row_spine to mz_repr::row (public) so
every consumer of Row's Columnar impl sees a uniform reference type.
Row's containers (Rows<BC, Vec<u8>>, borrowed Rows<BC, &[u8]>) now
produce DatumSeq<'a> on borrow / index, matching what DatumContainer
(the dictionary-compressed Row container used by RowRowSpine /
RowValSpine) already yields.

Consequences:

* Factorized arrangement cursors over Row-keyed spines now yield
  DatumSeq, which has the same Key<'a> type as DatumContainer-backed
  spines. This unblocks reduce_abelian's T2::Key<'a> = T1::Key<'a>
  bound, allowing reduce to emit FactLocal without a bridging
  arrangement.
* FactLocal closures in render/context.rs (as_collection, flat_map)
  switch from `|k: &RowRef|` to `|k: DatumSeq|`. `ToDatumIter` impls
  on both RowRef and DatumSeq keep external code working unchanged.
* DatumSeq gains a `Hash` impl (forwarded to byte-slice hash) so it
  works as a key for timely's columnar_exchange / logging hashes.
* DatumSeq gains `from_bytes` + `as_bytes` accessors (previously
  `as_bytes` was private to the compute crate).

No behavioral change; this is a type-system unification.
